backend/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
lightgbm
pandas
numpy
joblib
requests
python-dotenv
openai